Bunch of stuff (#70)
* Update hierarchy of folders in db
Add start and end for lp tasks and filter on it in the big query computing points
Create scripts to add new tasks for lp_points
Create new script to truncate only tables that we want
Add creation date & block on usg_markets
Update start => start_date and closed=> closed_date
Fix error in Onchain Tx bot that was canceling the notification to Telegram

* Add create_date and block for markets creation
Fix typo in renaming _date

* Fix linter

* Update reset_lp_points

* Fix truncate script

* Fix truncate

* Fix truncate

* Add logs

* Fix indexer block

* Fix boost seeding

* Update query with bigint to dont overflow
add console.log to debug date

* All the points in bigint in hell query

* Fix more int => bigint

* Fix add gauge script

---------

+++ src/scripts/utils/truncate_tables.ts
@@ -40,7 +40,8 @@ SELECT tablename AS table_name, schemaname AS schema_name FROM pg_tables WHERE s
 /**
  * Truncates all tables in the PostgreSQL database schema.
  */
-export async function truncateAllTables(): Promise<void> {
+export async function truncateTables(toDelete: string[]): Promise<void> {
+  const arr = toDelete.map((t) => t.toLowerCase())
   try {
     const confirmed = await promptConfirmation()
     if (!confirmed) {
@@ -62,10 +63,12 @@ export async function truncateAllTables(): Promise<void> {
 
     // Truncate each table
     for (const table of tables) {
-      const tableToTruncate = `"${table.schema_name}".${table.table_name}`
-      const truncateQuery = `TRUNCATE TABLE ${tableToTruncate} CASCADE`
-      await prisma.$executeRawUnsafe(truncateQuery)
-      console.log(`✅ Truncated: ${tableToTruncate}`)
+      if (arr.length === 0 || arr.includes(table.table_name)) {
+        const tableToTruncate = `"${table.schema_name}".${table.table_name}`
+        const truncateQuery = `TRUNCATE TABLE ${tableToTruncate} CASCADE`
+        await prisma.$executeRawUnsafe(truncateQuery)
+        console.log(`✅ Truncated: ${tableToTruncate}`)
+      }
     }
 
     // Re-enable foreign key constraints
@@ -78,6 +81,3 @@ export async function truncateAllTables(): Promise<void> {
     await prisma.$disconnect()
   }
 }
-
-// Execute the function
-truncateAllTables()
